+++ pons/_compiler.py
@@ -46,6 +46,9 @@ class EVMVersion(Enum):
     CANCUN = "cancun"
     """Cancun fork, Mar 13, 2024."""
 
+    PRAGUE = "prague"
+    """Prague fork, May 7, 2025."""
+
 
 def compile_contract_file(
     path: str | Path,